requirements.txt
fastapi
uvicorn
celery
redis
python-dotenv
groq
sentence-transformers
torch
numpy
pypdf
python-docx
